fix(#1208): avoid cross-suite URLProtocol mock race — separate mocks
