+++ Tests/AnglesiteRemoteTests/LoopbackHTTPExecutorTests.swift
@@ -3,6 +3,21 @@ import Foundation
 import AnglesiteP2P
 @testable import AnglesiteRemote
 
+final class MockURLProtocol: URLProtocol {
+    nonisolated(unsafe) static var handler: ((URLRequest) -> (HTTPURLResponse, Data))?
+
+    override class func canInit(with request: URLRequest) -> Bool { true }
+    override class func canonicalRequest(for request: URLRequest) -> URLRequest { request }
+    override func startLoading() {
+        guard let handler = Self.handler else { fatalError("no handler set") }
+        let (response, data) = handler(request)
+        client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
+        client?.urlProtocol(self, didLoad: data)
+        client?.urlProtocolDidFinishLoading(self)
+    }
+    override func stopLoading() {}
+}
+
 @Suite(.serialized) struct LoopbackHTTPExecutorTests {
     static func makeSession() -> URLSession {
         let config = URLSessionConfiguration.ephemeral
